Add Mixin and core dir
