+++ automator_mixins/_shuatu_base.py
@@ -1,0 +1,434 @@
+import time
+
+from core.cv import UIMatcher
+from ._base import BaseMixin
+
+
+class ShuatuBaseMixin(BaseMixin):
+    """
+    刷图基础插片
+    包含刷图基本操作和基本变量存储
+    """
+
+    def __init__(self):
+        super().__init__()
+        self.switch = 0
+        self.hard = 0  # 困难本出初始值
+        self.times = 3  # 总刷图次数
+
+    def sw_init(self):
+        self.switch = 0
+
+    def shuatuzuobiao(self, x, y, times):  # 刷图函数，xy为该图的坐标，times为刷图次数
+        if self.switch == 0:
+            tmp_cout = 0
+            self.d.click(x, y)
+            time.sleep(0.5)
+        else:
+            print('>>>无扫荡券或者无体力！', '结束 全部 刷图任务！<<<\r\n')
+            return
+        if self.switch == 0:
+            while True:  # 锁定加号
+                screen_shot_ = self.d.screenshot(format="opencv")
+                if UIMatcher.img_where(screen_shot_, 'img/jiahao.bmp'):
+                    # screen_shot = a.d.screenshot(format="opencv")
+                    for i in range(times - 1):  # 基础1次
+                        # 扫荡券不必使用opencv来识别，降低效率
+                        self.d.click(876, 334)
+                    time.sleep(0.3)
+                    self.d.click(758, 330)  # 使用扫荡券的位置 也可以用OpenCV但是效率不够而且不能自由设定次数
+                    time.sleep(0.3)
+                    screen_shot = self.d.screenshot(format="opencv")
+                    if UIMatcher.img_where(screen_shot, 'img/ok.bmp'):
+                        self.guochang(screen_shot, ['img/ok.bmp'], suiji=0)
+                    else:
+                        time.sleep(0.5)
+                        self.d.click(588, 370)
+                    # screen_shot = a.d.screenshot(format="opencv")
+                    # a.guochang(screen_shot,['img/shiyongsanzhang.jpg'])
+                    screen_shot_ = self.d.screenshot(format="opencv")
+                    if UIMatcher.img_where(screen_shot_, 'img/tilibuzu.jpg'):
+                        print('>>>无扫荡券或者无体力！结束此次刷图任务！<<<\r\n')
+                        self.switch = 1
+                        self.d.click(677, 458)  # 取消
+                        break
+                    screen_shot = self.d.screenshot(format="opencv")
+                    if UIMatcher.img_where(screen_shot, 'img/tiaoguo.jpg'):
+                        self.guochang(screen_shot, ['img/tiaoguo.jpg'], suiji=0)
+                        self.guochang(screen_shot, ['img/ok.bmp'], suiji=0)
+                    else:
+                        time.sleep(1)
+                        self.d.click(475, 481)  # 手动点击跳过
+                        self.guochang(screen_shot, ['img/ok.bmp'], suiji=0)
+                    break
+                else:
+                    if tmp_cout < 3:
+                        # 计时3次就失败
+                        self.d.click(x, y)
+                        time.sleep(0.5)
+                        tmp_cout = tmp_cout + 1
+                    else:
+                        print('>>>无扫荡券或者无体力！结束此次刷图任务！<<<\r\n')
+                        self.switch = 1
+                        self.d.click(677, 458)  # 取消
+                        break
+        else:
+            print('>>>无扫荡券或者无体力！结束刷图任务！<<<\r\n')
+        while True:
+            self.d.click(1, 1)
+            time.sleep(0.3)
+            screen_shot_ = self.d.screenshot(format="opencv")
+            if UIMatcher.img_where(screen_shot_, 'img/normal.jpg', at=(660, 72, 743, 94)):
+                break
+            if UIMatcher.img_where(screen_shot_, 'img/hard.jpg'):
+                break
+
+        # 进入hard图
+
+    def enterHardMap(self):
+        # 进入冒险
+        time.sleep(2)
+        self.d.click(480, 505)
+        time.sleep(2)
+        while True:
+            screen_shot_ = self.d.screenshot(format="opencv")
+            if UIMatcher.img_where(screen_shot_, 'img/dixiacheng.jpg'):
+                break
+        # 点击进入主线关卡
+        self.d.click(562, 253)
+        time.sleep(2)
+        while True:
+            screen_shot_ = self.d.screenshot(format="opencv")
+            if UIMatcher.img_where(screen_shot_, 'img/normal.jpg'):
+                self.d.click(880, 80)
+            if UIMatcher.img_where(screen_shot_, 'img/hard.jpg'):
+                break
+
+        # 左移动
+
+    def goLeft(self):
+        self.d.click(35, 275)
+        time.sleep(3)
+
+        # 右移动
+
+    def goRight(self):
+        self.d.click(925, 275)
+        time.sleep(3)
+        # 进入hard图
+
+    def goHardMap(self):
+        # 进入冒险
+        time.sleep(2)
+        self.d.click(480, 505)
+        time.sleep(2)
+        while True:
+            screen_shot_ = self.d.screenshot(format="opencv")
+            if UIMatcher.img_where(screen_shot_, 'img/dixiacheng.jpg'):
+                break
+        # 点击进入主线关卡
+        self.d.click(562, 253)
+        time.sleep(3)
+        self.d.click(828, 85)
+        time.sleep(2)
+        for _ in range(11):  # 设置大于当前进图数,让脚本能回归到1-1即可.
+            # n图左移到1-1图
+            time.sleep(2)
+            self.d.click(27, 272)
+        while True:
+            screen_shot_ = self.d.screenshot(format="opencv")
+            if UIMatcher.img_where(screen_shot_, 'img/normal.jpg'):
+                self.d.click(828, 85)
+            else:
+                UIMatcher.img_where(screen_shot_, 'img/hard.jpg')
+                break
+
+    def hard_shuatuzuobiao(self, x, y, times):  # 刷图函数，xy为该图的坐标，times为刷图次数,防止占用shuatuzuobiao用的
+        if self.switch == 0:
+            tmp_cout = 0
+            self.d.click(x, y)
+            time.sleep(0.5)
+        else:
+            print('>>>无扫荡券或者无体力！', '结束 全部 刷图任务！<<<\r\n')
+            return
+        if self.switch == 0:
+            while True:  # 锁定加号
+                screen_shot_ = self.d.screenshot(format="opencv")
+                if UIMatcher.img_where(screen_shot_, 'img/jiahao.bmp'):
+                    # screen_shot = a.d.screenshot(format="opencv")
+                    for i in range(times - 1):  # 基础1次
+                        # 扫荡券不必使用opencv来识别，降低效率
+                        self.d.click(876, 334)
+                    time.sleep(0.3)
+                    self.d.click(758, 330)  # 使用扫荡券的位置 也可以用OpenCV但是效率不够而且不能自由设定次数
+                    time.sleep(0.3)
+                    screen_shot = self.d.screenshot(format="opencv")
+                    if UIMatcher.img_where(screen_shot, 'img/ok.bmp'):
+                        self.guochang(screen_shot, ['img/ok.bmp'], suiji=0)
+                    else:
+                        time.sleep(0.5)
+                        self.d.click(588, 370)
+                    # screen_shot = a.d.screenshot(format="opencv")
+                    # a.guochang(screen_shot,['img/shiyongsanzhang.jpg'])
+                    screen_shot_ = self.d.screenshot(format="opencv")
+                    if UIMatcher.img_where(screen_shot_, 'img/tilibuzu.jpg'):
+                        print('>>>无扫荡券或者无体力！结束此次刷图任务！<<<\r\n')
+                        self.switch = 1
+                        self.d.click(677, 458)  # 取消
+                        break
+                    screen_shot = self.d.screenshot(format="opencv")
+                    if UIMatcher.img_where(screen_shot, 'img/tiaoguo.jpg'):
+                        self.guochang(screen_shot, ['img/tiaoguo.jpg'], suiji=0)
+                        self.guochang(screen_shot, ['img/ok.bmp'], suiji=0)
+                    else:
+                        time.sleep(1)
+                        self.d.click(475, 481)  # 手动点击跳过
+                        self.guochang(screen_shot, ['img/ok.bmp'], suiji=0)
+                    break
+                else:
+                    if tmp_cout < 3:
+                        # 计时3次就失败
+                        self.d.click(x, y)
+                        time.sleep(0.5)
+                        tmp_cout = tmp_cout + 1
+                    else:
+                        print('>>>无扫荡券或者无体力！结束此次刷图任务！<<<\r\n')
+                        self.switch = 1
+                        self.d.click(677, 458)  # 取消
+                        break
+        else:
+            print('>>>无扫荡券或者无体力！结束刷图任务！<<<\r\n')
+        while True:
+            self.d.click(1, 1)
+            time.sleep(0.3)
+            screen_shot_ = self.d.screenshot(format="opencv")
+            if UIMatcher.img_where(screen_shot_, 'img/normal.jpg', at=(660, 72, 743, 94)):
+                break
+            if UIMatcher.img_where(screen_shot_, 'img/hard.jpg'):
+                break
+
+    def ziduan00(self):
+        print('>>>识别到00参数该字段不刷图,要刷图请更改zhanghao.txt！结束刷图任务！<<<\r\n')
+        self.lockimg('img/liwu.bmp', elseclick=[(131, 533)], elsedelay=1, at=(891, 413, 930, 452))  # 回首页
+
+    # 继续执行函数
+    def continueDo9(self, x, y):
+        self.switch = 0
+        self.shuatuzuobiao(x, y, self.times)  # 3-3
+
+    # 刷活动hard图
+    def doActivityHard(self):
+        # 进入冒险
+        time.sleep(2)
+        self.d.click(480, 505)
+        time.sleep(2)
+        while True:
+            screen_shot_ = self.d.screenshot(format="opencv")
+            if UIMatcher.img_where(screen_shot_, 'img/dixiacheng.jpg'):
+                break
+        # 点击进入活动
+        self.d.click(415, 430)
+        time.sleep(3)
+        while True:
+            screen_shot_ = self.d.screenshot(format="opencv")
+            self.d.click(480, 380)
+            time.sleep(0.5)
+            self.d.click(480, 380)
+            if UIMatcher.img_where(screen_shot_, 'img/normal.jpg'):
+                self.d.click(880, 80)
+            if UIMatcher.img_where(screen_shot_, 'img/hard.jpg'):
+                break
+        self.shuatuzuobiao(689, 263, self.times)  # 1-5
+        self.continueDo9(570, 354)  # 1-4
+        self.continueDo9(440, 255)  # 1-3
+        self.continueDo9(300, 339)  # 1-2
+        self.continueDo9(142, 267)  # 1-1
+        self.lockimg('img/liwu.bmp', elseclick=[(131, 533)], elsedelay=1, at=(891, 413, 930, 452))  # 回首页
+
+        # 识别7村断崖
+
+    def duanyazuobiao(self):
+        self.tag = 0
+        time.sleep(2)
+        while True:
+            screen_shot_ = self.d.screenshot(format="opencv")
+            if UIMatcher.img_where(screen_shot_, 'img/duanyazuobiao.jpg'):
+                print('>>>成功识别标记,开始刷图.<<<\r\n')
+                break
+            if self.tag > 11:  # 超过11次点击则不刷图
+                print('>>>点击超过11次,已错过标记点！刷图任务异常,开启高延迟！<<<\r\n')
+                print('>>>提示:请增加duanyazuobiao的延迟时间<<<\r\n')
+                if UIMatcher.img_where(screen_shot_, 'img/duanyazuobiao.jpg'):
+                    break
+                if self.tag > 22:  # 超过11次点击则不刷图
+                    print('>>>已错过标记点,高延迟模式失败,放弃刷图.<<<\r\n')
+                    print('>>>提示:请增加duanyazuobiao的延迟时间<<<\r\n')
+                    break
+                else:
+                    for _ in range(1):
+                        self.d.click(925, 275)
+                        self.tag += 1
+                        time.sleep(4)  # 这是高延迟识别时间,模拟器卡顿请加时.
+            else:
+                for _ in range(1):
+                    self.d.click(27, 272)
+                    self.tag += 1
+                    time.sleep(2)  # 这是延迟识别时间,模拟器卡顿请加时.
+
+    def shoushuazuobiao(self, x, y, jiaocheng=0, lockpic='img/normal.jpg', screencut=None):
+        """
+        不使用挑战券挑战，xy为该图坐标
+        jiaocheng=0 只处理简单的下一步和解锁内容
+        jiaocheng=1 要处理复杂的教程
+        lockpic: 返回时锁定的图
+        screencut: 返回时锁定的图的搜索范围
+        :return:
+        """
+        while True:
+            screen_shot_ = self.d.screenshot(format="opencv")
+            if UIMatcher.img_where(screen_shot_, lockpic, at=screencut):
+                break
+            self.d.click(1, 138)
+            time.sleep(1)
+        self.lockimg('img/tiaozhan.jpg', elseclick=[(x, y)], elsedelay=2)
+        self.d.click(840, 454)
+        time.sleep(0.7)
+
+        while True:
+            screen_shot_ = self.d.screenshot(format="opencv")
+            if UIMatcher.imgs_where(screen_shot_, ['img/kuaijin.jpg', 'img/kuaijin_1.jpg']) != {}:
+                break
+            self.d.click(840, 454)  # 点到进入战斗画面
+            time.sleep(0.7)
+        while True:
+            screen_shot_ = self.d.screenshot(format="opencv")
+            if self.click_img(screen_shot_, 'img/kuaijin.jpg', at=(891, 478, 936, 517)):
+                time.sleep(1)
+            if self.click_img(screen_shot_, 'img/auto.jpg', at=(891, 410, 936, 438)):
+                time.sleep(1)
+            if UIMatcher.img_where(screen_shot_, 'img/wanjiadengji.jpg', at=(233, 168, 340, 194)):
+                break
+            self.d.click(1, 138)
+            time.sleep(0.5)
+        if jiaocheng == 1:  # 有复杂的教程，交给教程函数处理
+            self.chulijiaocheng()
+        else:  # 无复杂的教程，自己处理掉“下一步”
+            for _ in range(7):
+                self.d.click(832, 506)
+                time.sleep(0.2)
+            while True:
+                time.sleep(2)
+                screen_shot_ = self.d.screenshot(format="opencv")
+                if UIMatcher.img_where(screen_shot_, lockpic, at=screencut):
+                    break
+                elif UIMatcher.img_where(screen_shot_, 'img/xiayibu.jpg'):
+                    self.d.click(832, 506)
+                else:
+                    self.d.click(1, 100)
+            while True:  # 两次确认回到挑战界面
+                self.d.click(1, 100)
+                time.sleep(0.5)
+                screen_shot_ = self.d.screenshot(format="opencv")
+                if UIMatcher.img_where(screen_shot_, lockpic, at=screencut):
+                    break
+
+    def qianghua(self):
+        # 此处逻辑极为复杂，代码不好理解
+        time.sleep(3)
+        self.d.click(215, 513)  # 角色
+        time.sleep(3)
+        self.d.click(177, 145)  # First
+        time.sleep(3)
+        for i in range(5):
+            print("Now: ", i)
+            while True:
+                screen_shot_ = self.d.screenshot(format='opencv')
+                if UIMatcher.img_where(screen_shot_, 'img/keyihuode.jpg'):
+                    # 存在可以获得，则一直获得到没有可以获得，或者没有三星
+                    self.d.click(374, 435)
+                    time.sleep(1)
+                    screen_shot_ = self.d.screenshot(format='opencv')
+                    if UIMatcher.img_where(screen_shot_, 'img/tuijianguanqia.jpg', at=(258, 87, 354, 107)):
+                        # 已经强化到最大等级，开始获取装备
+                        if not UIMatcher.img_where(screen_shot_, 'img/sanxingtongguan.jpg'):
+                            # 装备不可刷，换人
+                            self.d.click(501, 468)  # important
+                            time.sleep(1)
+                            break
+                        while UIMatcher.img_where(screen_shot_, 'img/sanxingtongguan.jpg'):
+                            # 一直刷到没有有推荐关卡但没有三星或者返回到角色列表
+                            self.guochang(screen_shot_, ['img/sanxingtongguan.jpg'], suiji=0)
+                            time.sleep(1)
+                            # 使用扫荡券的数量：
+                            for _ in range(4 - 1):
+                                self.d.click(877, 333)
+                                time.sleep(0.3)
+                            self.d.click(752, 333)
+                            time.sleep(0.7)
+                            self.d.click(589, 371)
+                            while True:
+                                screen_shot_ = self.d.screenshot(format='opencv')
+                                active_paths = UIMatcher.imgs_where(screen_shot_,
+                                                                    ['img/tuijianguanqia.jpg', 'img/zidongqianghua.jpg',
+                                                                     'img/tiaoguo.jpg'])
+                                if 'img/tiaoguo.jpg' in active_paths:
+                                    x, y = active_paths['img/tiaoguo.jpg']
+                                    self.d.click(x, y)
+                                if 'img/tuijianguanqia.jpg' in active_paths:
+                                    flag = 'img/tuijianguanqia.jpg'
+                                    break
+                                elif 'img/zidongqianghua.jpg' in active_paths:
+                                    flag = 'img/zidongqianghua.jpg'
+                                    break
+                                else:
+                                    self.d.click(1, 100)
+                                    time.sleep(1.3)
+                            if flag == 'img/zidongqianghua.jpg':
+                                # 装备获取完成，跳出小循环，重进大循环
+                                self.d.click(371, 437)
+                                time.sleep(0.7)
+                                break
+                            else:
+                                # 装备未获取完毕，继续尝试获取
+                                continue
+                        self.d.click(501, 468)  # important
+                        time.sleep(2)
+                        continue
+                    else:
+                        # 未强化到最大等级，强化到最大登记
+                        self.d.click(501, 468)  # important
+                        time.sleep(3)
+                        continue
+                else:
+                    # 没有可以获得
+                    if UIMatcher.img_where(screen_shot_, 'img/ranktisheng.jpg', at=(206, 325, 292, 346)):
+                        self.d.click(250, 338)
+                        time.sleep(2)
+                        screen_shot_ = self.d.screenshot(format='opencv')
+                        active_list = UIMatcher.imgs_where(screen_shot_, ['img/queren.jpg', 'img/ok.bmp'])
+                        if 'img/queren.jpg' in active_list:
+                            x, y = active_list['img/queren.jpg']
+                            self.d.click(x, y)
+                        if 'img/ok.bmp' in active_list:
+                            x, y = active_list['img/ok.bmp']
+                            self.d.click(x, y)
+                        time.sleep(8)
+                        self.d.click(481, 369)
+                        time.sleep(1)
+                        continue
+                    else:
+                        self.d.click(371, 437)
+                        time.sleep(0.7)
+                        self.d.click(501, 468)  # important
+                        time.sleep(2)
+                        break
+            self.d.click(933, 267)  # 下一位
+            time.sleep(2)
+
+        self.lockimg('img/liwu.bmp', elseclick=[(131, 533)], elsedelay=1, at=(891, 413, 930, 452))  # 回首页
+        self.lockimg('img/zhuxianguanqia.jpg', elseclick=[(480, 513)], elsedelay=3)
+        self.d.click(562, 253)
+        time.sleep(3)
+        self.lockimg('img/normal.jpg', elseclick=[(704, 84)], elsedelay=0.5, alldelay=1, at=(660, 72, 743, 94))
+        self.d.click(923, 272)
+        time.sleep(3)
