+++ core/constant.py
@@ -1,0 +1,11 @@
+MAIN_BTN = {
+    "hanghui": (693, 430),  # 行会按钮
+
+}
+HANGHUI_BTN = {
+    "chengyuan": (241, 350),  # 点击成员
+
+}
+
+# 显然后面还没写
+# 等有空补全
