+++ gettili.py
@@ -1,16 +1,15 @@
 # coding=utf-8
-import uiautomator2 as u2
-import time
-from utils import *
-from cv import *
-from Automator import *
 # import matplotlib.pylab as plt
 import os
 import threading
-import log_handler
 
-acclog=log_handler.pcr_acc_log()
-#测试程序
+from core import log_handler
+from core.Automator import *
+
+acclog = log_handler.pcr_acc_log()
+
+
+# 测试程序
 
 def runmain(address, account, password):
     # 主功能体函数
@@ -19,27 +18,26 @@ def runmain(address, account, password):
     a = Automator(address, account)
     a.start()
 
-
     # #opencv识别可视化 无法在多线程中使用
     # plt.ion()
     # fig, ax = plt.subplots(1)
     # plt.show()
 
-    #print('>>>>>>>即将登陆的账号为：', account, '密码：', password, '<<<<<<<')
+    # print('>>>>>>>即将登陆的账号为：', account, '密码：', password, '<<<<<<<')
 
-    a.log.write_log(level='info',message='>>>>>>>即将登陆的账号为：%s 密码：%s<<<<<<<'%(account,password))
+    a.log.write_log(level='info', message='>>>>>>>即将登陆的账号为：%s 密码：%s<<<<<<<' % (account, password))
 
     acclog.Account_Login(account)
     a.login_auth(account, password)  # 注意！请把账号密码写在zhanghao2.txt内
     a.init_home()  # 初始化，确保进入首页
 
-    #a.tansuo()  # 探索
-    #a.dixiachengDuanya()  # 地下城，请把队伍列表里1队设置为打boss队，2队设置为aoe队
-    #a.shouqurenwu()  # 收取任务
-    #a.shouqu()  # 收取所有礼物
-    #print(a.baidu_ocr(0,0,200,200,2.0))
+    # a.tansuo()  # 探索
+    # a.dixiachengDuanya()  # 地下城，请把队伍列表里1队设置为打boss队，2队设置为aoe队
+    # a.shouqurenwu()  # 收取任务
+    # a.shouqu()  # 收取所有礼物
+    # print(a.baidu_ocr(0,0,200,200,2.0))
 
-    #print('当前体力值为：',a.get_tili())
+    # print('当前体力值为：',a.get_tili())
     acclog.Account_Logout(account)
     a.change_acc()  # 退出当前账号，切换下一个
